trend_discovery.py
import requests
from bs4 import BeautifulSoup
import logging
import random
import time

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

HEADERS = {
    'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
}

def get_trending_tweet_patterns():
    """Discover trending tweet patterns (mock implementation)"""
    
    # Since direct Twitter scraping is complex and may violate ToS,
    # this returns curated trending patterns based on current social media trends
    
    trending_patterns = [
        "Hook + numbered list format",
        "Question-based engagement tweets",
        "Personal story with lesson learned",
        "Contrarian takes on popular topics",
        "Behind-the-scenes content",
        "Thread starters with cliffhangers",
        "Relatable daily struggles",
        "Quick tips and hacks",
        "Motivational morning thoughts",
        "Weekend reflection posts"
    ]
    
    # Simulate API delay
    time.sleep(random.uniform(1, 3))
    
    # Return random selection of patterns
    selected_patterns = random.sample(trending_patterns, 5)
    logger.info(f"Discovered trending patterns: {selected_patterns}")
    
    return selected_patterns

def get_hashtag_trends():
    """Get trending hashtags (mock implementation)"""
    trending_hashtags = [
        "#MondayMotivation", "#TechTips", "#ProductivityHack",
        "#WeekendReflections", "#StartupLife", "#RemoteWork",
        "#LifeLessons", "#GrowthMindset", "#Innovation", "#Success"
    ]
    
    return random.sample(trending_hashtags, 3)