requirements.txt
requests
tweepy
schedule
pandas
python-dotenv
beautifulsoup4
google-generativeai
